Rename package "wareply" to "watomatic"
