+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion 30
 
     defaultConfig {
-        applicationId "com.parishod.wareply"
+        applicationId "com.parishod.watomatic"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
