+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.parishod.wareply">
+    package="com.parishod.watomatic">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.WaReply">
 
-        <activity android:name=".activity.main.MainActivity">
+        <activity android:name="com.parishod.watomatic.activity.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -19,20 +19,20 @@
         </activity>
 
         <activity
-            android:name=".activity.customreplyeditor.CustomReplyEditorActivity"
+            android:name="com.parishod.watomatic.activity.customreplyeditor.CustomReplyEditorActivity"
             android:label="@string/mainAutoReplyLabel"
-            android:parentActivityName=".activity.main.MainActivity">
+            android:parentActivityName="com.parishod.watomatic.activity.main.MainActivity">
         </activity>
 
         <activity
-            android:name=".activity.about.AboutActivity"
+            android:name="com.parishod.watomatic.activity.about.AboutActivity"
             android:label="@string/about"
-            android:parentActivityName=".activity.main.MainActivity"
+            android:parentActivityName="com.parishod.watomatic.activity.main.MainActivity"
             >
         </activity>
 
         <service
-            android:name=".NotificationService"
+            android:name="com.parishod.watomatic.NotificationService"
             android:label="@string/service_name"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE">
             <intent-filter>
