+++ app/src/main/java/com/parishod/watomatic/activity/customreplyeditor/CustomReplyEditorActivity.java
@@ -1,4 +1,4 @@
-package com.parishod.wareply.activity.customreplyeditor;
+package com.parishod.watomatic.activity.customreplyeditor;
 
 import android.os.Bundle;
 import android.text.Editable;
@@ -8,8 +8,8 @@ import android.widget.Button;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.material.textfield.TextInputEditText;
-import com.parishod.wareply.R;
-import com.parishod.wareply.model.CustomRepliesData;
+import com.parishod.watomatic.R;
+import com.parishod.watomatic.model.CustomRepliesData;
 
 public class CustomReplyEditorActivity extends AppCompatActivity {
     TextInputEditText autoReplyText;
