+++ app/src/main/java/com/parishod/watomatic/model/logs/whatsapp/WhatsappAutoReplyLogsDao.java
@@ -1,4 +1,4 @@
-package com.parishod.wareply.model.logs.whatsapp;
+package com.parishod.watomatic.model.logs.whatsapp;
 
 import androidx.room.Dao;
 import androidx.room.Insert;
